+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## [Unreleased]
 
+### added
+
+- Added the `--recursive` argument to the `init` subcommand
+- Added the `--recursive` argument to the `update` subcommand
+
+### removed
+
+- Removed the `--cascade` argument from the `approve` subcommand
+- Removed the `--cascade` argument from the `reject` subcommand
+
 ### changed
 
 - Renamed the subcommands to be more simple example: `init-courses` to `init`